Fix the SimDate command format
The plugin was using an undocumented single argument version of the
command that used the game's internal date format as a workaround for
the next day bug in the MM DD YYYY version.
The single argument command no longer works for some reason, so we are
now using the MM DD YYYY version.

The current date is now stored as a clone of the cIGZDate instance we
retrieve when exiting a city.
This allows us to use the game's code for subtracting days and
comparing cIGZDate instances.
Subtracting days is required due to a bug in the SimDate command, it
set the city date to the next day after the one you entered.